Move filtering of excluded nodes to the spanner implementation, and update of handling of empty responses to still include the node (with an empty response).
