+++ internal/server/v2/bulk_variable_group_info/bulk_variable_group_info_hidden_excluded.json
@@ -1,1 +1,8 @@
-{}
+{
+  "data": [
+    {
+      "node": "dc/g/Hidden",
+      "info": {}
+    }
+  ]
+}
